tests/inputdevice-keyboard.html

<!DOCTYPE html>
<title>InputDevice test for keyboard-triggered events</title>
<meta name="viewport" content="width=device-width">
<script src="https://w3c-test.org/resources/testharness.js"></script>
<script src="https://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for keyboard-triggered events</h1>
<p>Tab into the input fields below, enter some text and activate the checkbox using nothing but a keyboard.</p>
<div id="target" tabindex=1>
  <input>
  <input type=checkbox>
</div>

<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(event.sourceDevice != null, "sourceDevice is set");
  assert_false(event.sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function runTests() {
  var events = ["click", "focus", "focusin", "focusout", "blur", "keydown", "keyup", "keypress"];
  // Some browsers don't yet implement 'input' as a UIEvent, and many browsers don't yet support
  // 'beforeinput'
  if ('InputEvent' in window)
    events = events.join(["input", "beforeinput"]);
  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
